Libft complete. Makefile not finish

+++ ft_lstsize.c
@@ -6,7 +6,7 @@
 /*   By: bwisniew <bwisniew@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/09 16:13:01 by bwisniew          #+#    #+#             */
-/*   Updated: 2023/11/09 16:17:49 by bwisniew         ###   ########.fr       */
+/*   Updated: 2023/11/10 10:12:32 by bwisniew         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	size_t	size;
+	int	size;
 
 	size = 0;
 	while (lst != 0)
